fix: check all commits for the commit_date, not just a range
Previously, the script stopped when it encountered a commit_date
that was before the target date, it stopped checking all other commits,
but it's possible that the commits are not listed in chronological
order, such as in this case:

* e7caea7 2025-04-26 feat(gld): delete config
* 0d283fa 2025-03-14 feat: [wip] add sqlite_db_backup script
* f41316d 2025-04-26 feat: [wip] add updated system-actions

Now we check all commits. The alternative would be to sort them before
checking, but it's hardly more performant.